+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file. The changes
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 10.3.0-[Unreleased] - ???
+## 10.4.0-[Unreleased] - ???
 
 [Artifacts](https://nexus3.inventage.com/#browse/search=version%3D%22???%22)
+
+## [10.3.0]-202602121511-92-c7ee225 - 2026-02-12
+
+[Artifacts](https://nexus3.inventage.com/#browse/search=version%3D%2210.3.0-202602121511-92-c7ee225%22)
 
 ### Added
 
@@ -701,7 +705,8 @@ Starting with this release, please use the following OCI registry `uniportcr.art
 
 - Uniport-Gateway providing reverse proxy functionality within the Inventage Portal Solution ([PORTAL-89](https://inventage-all.atlassian.net/browse/PORTAL-89)).
 
-[unreleased]: https://github.com/uniport/uniport-gateway/compare/10.2.0...main
+[unreleased]: https://github.com/uniport/uniport-gateway/compare/10.3.0...main
+[10.3.0]: https://github.com/uniport/uniport-gateway/compare/10.2.0...10.3.0
 [10.2.0]: https://github.com/uniport/uniport-gateway/compare/10.1.0...10.2.0
 [10.1.0]: https://github.com/uniport/uniport-gateway/compare/10.0.0...10.1.0
 [10.0.0]: https://github.com/uniport/uniport-gateway/compare/9.5.0...10.0.0